requirements-lock.txt
alabaster==0.7.12
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
docutils==0.16
idna==2.9
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.1
keyring==21.1.1
lazy-object-proxy==1.4.3
livereload==2.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
natsort==6.2.1
ordered-set==3.1.1
packaging==20.3
pathlib2==2.3.5
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.8.1
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
pywin32-ctypes==0.2.0
pyxcute==0.6.0
readme-renderer==24.0
requests==2.23.0
requests-toolbelt==0.9.1
semver==2.9.1
Send2Trash==1.5.0
setuptools==41.2.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tornado==6.0.4
tqdm==4.43.0
twine==3.1.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
wheel==0.34.2
wrapt==1.11.2